backend/requirements.txt
# Backend requirements for FastAPI and geospatial processing
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic>=1.10,<2.0
geopy
httpx
